+++ src/main/java/io/github/ash/fastcrops/FastCropsPlugin.java
@@ -4,6 +4,7 @@ import io.github.ash.fastcrops.command.FastCropsCommand;
 import io.github.ash.fastcrops.config.FastCropsConfig;
 import io.github.ash.fastcrops.growth.GrowthEngine;
 import io.github.ash.fastcrops.tracking.GrowableTracker;
+import io.github.ash.fastcrops.tripwire.TripwireExtensionService;
 import io.github.ash.fastcrops.update.AutoUpdater;
 import org.bukkit.command.PluginCommand;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -12,6 +13,7 @@ public final class FastCropsPlugin extends JavaPlugin {
     private FastCropsConfig fastCropsConfig;
     private GrowableTracker growableTracker;
     private GrowthEngine growthEngine;
+    private TripwireExtensionService tripwireExtensionService;
     private AutoUpdater autoUpdater;
 
     @Override
@@ -24,6 +26,9 @@ public final class FastCropsPlugin extends JavaPlugin {
         this.growableTracker = new GrowableTracker(this, fastCropsConfig);
         getServer().getPluginManager().registerEvents(growableTracker, this);
         this.growableTracker.rebuildFromLoadedChunks();
+
+        this.tripwireExtensionService = new TripwireExtensionService(this, fastCropsConfig);
+        getServer().getPluginManager().registerEvents(tripwireExtensionService, this);
 
         this.growthEngine = new GrowthEngine(this, fastCropsConfig, growableTracker);
         this.growthEngine.start();